+++ src/pages/common/_chantest_page.c
@@ -14,7 +14,7 @@
  */
 static struct chantest_page * const cp = &pagemem.u.chantest_page;
 static struct chantest_obj * const gui = &gui_objs.u.chantest;
-static s16 showchan_cb(void *data);
+static s32 showchan_cb(void *data);
 static const char *value_cb(guiObject_t *obj, const void *data);
 static const char *channum_cb(guiObject_t *obj, const void *data);
 static void _handle_button_test();
@@ -72,7 +72,7 @@ void PAGE_ChantestEvent()
         _handle_button_test();
         return;
     }
-    volatile s16 *raw = MIXER_GetInputs();
+    volatile s32 *raw = MIXER_GetInputs();
     for(int i = 0; i < cp->num_bars; i++) {
         int ch = get_channel_idx(cur_row * NUM_BARS_PER_ROW + i);
         int v = RANGE_TO_PCT((ch >= NUM_INPUTS && ch < NUM_INPUTS + NUM_OUT_CHANNELS)
@@ -93,7 +93,7 @@ void PAGE_ChantestExit()
 {
     BUTTON_UnregisterCallback(&cp->action);
 }
-static s16 showchan_cb(void *data)
+static s32 showchan_cb(void *data)
 {
     long ch = (long)data;
     return cp->pctvalue[ch];
